Make test module to export results in xml
Refactor test module to export the results in xml and make use of the xml
representation in OpenQA.

As a PoC i have modified only one module for now. But i am planning to refactor
other (HPC) modules as such. That's why the main functions are implemented into
`utils.pm` so everyone can take advantage of this format, if he/she wants to.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>