Implement new module to test apptainer
Apptainer is missing testing in both SLE and OpenSUSE but the package is
available only in OpenSUSE. The module tests some basic functionalities of the
tool to pull, build and run containers.

One of the things left out is the `capability` command which seems to not work
properly at the moment.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>