Add repositories to both machines and force update
When we want to test against repositories like science:hpc we should
a) enable this repository on both machines
and
b) make sure we switch packages over to the ones provided there.

Only this way we can be sure we are testing the complete stack provided
from this repository. Plus, installation of packages fails with
unresolved dependencies otherwise.

Signed-off-by: Michael Moese <mmoese@suse.com>