Remove validation from svirt-hyperv failing jobs
We need to remove the validations from jobs like svirt-hyperv in yast dev,
cause this kind of virtualization is pretty slow and the problem is not
going to be fixed any time soon.

And we leave msdos@svirt-hyperv in the yast dev job goup to label this kind
of error for issue tracking.