Boot JeOS vmware image configured with cloud-init
Starting with QR1 of SLE15sp4, VMWare images can be configure either by
jeos-firstboot or cloud-init. Therefore we need to re-use `FIRST_BOOT_CONFIG`
variable in order to determine which tool is expected in the test.

This PR also updates the cloud-init data file as former localization
workaround is not required any more.

- ticket: [Test VMware with CI](https://progress.opensuse.org/issues/117409)