Don't install packages during system installation
Currently, we install rdma-core and rdma-ndd during system installation.
Initially the idea was to save a reboot after installation, but in the
meantime, we both install these packages and do an unconditional reboot
in ibtests_prepare.pm. So it is perfectly fine to remove this from our
schedules, and thus avoid failing needle checks during installation.

Signed-off-by: Michael Moese <mmoese@suse.com>