Select last item on select_patterns
Until now the workaround did not select the last item because the up
button was pressed last. With this new fix we will only send up when the
screen did not change after we have typed the second down, which means
that we are on the last item. This will result with the last item being
selected.