Install gnu-hpc packages on HPC compute nodes
Due to the `openmpi` updates the mpi tests have been refactor. Those changes
breaks the tests which uses `openmpi` implementation. As for now the
spack test suite uses *mpich* so the test suite is not affected directly.
However, with this refactoring, we remove the use of `get_compute_nodes_deps`
which is now deprecated. The only binaries shared with NFS is the directory
with the code to run and `prepare_spack_env` takes care of the installation
of the required libraries

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>