wicked: Delay ifup of t08_setup_second_card
OVS is blocking the port in the beginning sometimes. This cause trouble
in this test, as the first DHCP-Request times out and wicked is
using the last lease (which is something from the installation).

Now we simply wait 30s to give OVS some time. Most properly this is
something related to RSTP configured on the OVS bridge.

Reference: bsc#1204604