Update mpirun parameters to prevent warnings when in Ethernet
With recent updates the `openmpi` throws warnings when mpirun runs. That should
occur when the Ethernet is used instead of the Infiniband. AFAIK the detection
of one or the other is not easy task and even enabling a whole bunch of log it's
not easy to be sure what is really being used

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>