Skip extra_tests_phub test if PackageHub is not ready
https://progress.opensuse.org/issues/120879

We can skip all tests which need PackageHub activated via parameter
"PHUB_REAY=0". at the same time, it might be not ready in early
development phase, so if we are not sure it is ready, we can have a try
first to activate it and then force fail the job if it is not available.