Add AMI for us-east-1
us-east-1 is used for our arm64 test runs, however until now we always
had to specify the PUBLIC_CLOUD_EC2_UPLOAD_AMI setting there manually.
This commit adds an entry for a static AMI for usage in this region, so
that the PUBLIC_CLOUD_EC2_UPLOAD_AMI setting is not needed anymore
unconditionally.