Add AWS configuration when running on EC2 and wait for ssh
Regression tests of https://github.com/SUSE/qe-sap-deployment in openQA
are currently failing on EC2, because the ansible commands over SSH are
called before the cloud infrastructure is ready for SSH connections, and
because of a missing `~/.aws/config` file referenced by the ansible
playbooks which is not present in the jumphost.

This commit fixes both issues by adding new methods into lib/qesapdeployment.pm
to create the configuration files required by AWS/EC2 and to get the public IP
addresses from the terraform deployment so the module can wait until SSH is ready.