Provide workaround for iscsi failure bsc#1207157
Iscsi MU tests for 15 SP3 are failing due to yast2-iscsi-client return
with an unexpected exit code. This is an issue that has already appeared
on 15 SP5, was addressed with bsc#1206132 and subsequently fixed.
On 15 SP3, while waiting for the fix to land, in order to not block
maintainance updates due to failure we apply the fix manually, as
workaround.
The workaround is pretty straightforward and it requires to edit the
iscsid service file.