Fix the hibernation_enabled issue on aarch64
See https://progress.opensuse.org/issues/123394.
Previous builds use hibernation_disabled but now we switched to hibernation_enabled and this might be a problem, because in the default filesystem option the checkmark for "Enlarge to RAM size" for the swap partition is not set on aarch64. A question about this was answered by Josef Reidinger in the help-yast Slack channel, the default is only set for x86_64, so no hibernation is enabled on aarch. That means we cannot use the same schedule for guided_btrfs on both aarch64 and x86_64.