Update the Trento tool image generator
Update the schedule to generate the Trento tool qcow2 image.
The new sequence is now the same of the PC tool image, with just the
Trento specific step to install some tools  on top of it.
Add multiple tag support: TRENTO_CYPRESS_VERSION is now an array var
Code for the test and lib simplified