Fix mismatched parameter error in boot parameters
On specific tests the bootloader parameters check finds a mismatch
between the expected and received boot params.
This happens because, while all the parameters have been typed in the
boot options command line in the grub menu (max char length is 512)
only 477 of those characters are recorded, if you check the
/proc/cmdline of the system (you could check it during installation).
The reason is that the system will automatically add the initrd and
splash parameters, and even then, /proc/cmdline only allows for 506
characters.
Given these findings the only way to avoid characters being cropped,
due to exceeding the 477 limit, is to shorten the urls added as
parameters. Since the system installation complains if the self_update
url is shortened, the solution was to shorten the dud url. This solution
makes sense because it seems to be the introduction of the dud url
parameter that causes this mismatch compared to other tests that do not
display parameter mismatch.